packer/environment/python/requirements.txt
requests
docker
python-decouple
